add summ Number
