+++ src/pages/MainPage.js
@@ -3,8 +3,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 //Components
 //Тимчасово добавлю сюди так як немає навігації
-import Card from '../components/Card/Card';
-import DaysList from '../components/DaysList/DaysList';
+import Card from '../components/Commons/Card/Card';
+import DaysList from '../components/Commons/DaysList/DaysList';
 
 const MainPage = () => {
 	return (
